Update Pocket Plane behavior while quest is active
Originally this mod replaced the Pocket Plane abilility with a fake version
while the quest was active which displayed a message that the ability was not
available. However, the game seems to have hardcoded the originally PP ability,
so that it could not be removed permanently from the protagonist.

This update allows to use the PP ability but makes sure that you can't really
leave the confinements of the castle while the quest is active.